moved roi_stats to a new stats.py module
