+++ suite2p/io/nwb.py
@@ -4,8 +4,7 @@ import os
 import numpy as np
 import scipy
 
-from .. import detection
-from .. import extraction
+from ..detection import roi_stats
 from .. import run_s2p
 
 try:
@@ -43,7 +42,7 @@ def read_nwb(fpath):
             if multiplane:
                 stat[-1]['iplane'] = int(rois[n][0,-2])
         ops = run_s2p.default_ops()
-        stat = detection.roi_stats(ops, stat)
+        stat = roi_stats(ops, stat)
         if multiplane:
             nplanes = np.max(np.array([stat[n]['iplane'] for n in range(len(stat))]))+1
         else:
